Do day 6 part 2

+++ src/day6part1.rs
@@ -5,7 +5,7 @@ pub fn solve() {
     println!("{}", process_lines(lines));
 }
 
-fn process_lines(lines: Vec<String>) -> u32 {
+fn process_lines(lines: Vec<String>) -> u64 {
     let mut multiplication = 1;
 
     let times = get_numbers_from_string_with_colon(lines[0].as_str());
@@ -16,13 +16,13 @@ fn process_lines(lines: Vec<String>) -> u32 {
         let different_tries = times[i] / 2;
 
         for j in 1..different_tries {
-            if does_new_held_time_beat_record(times[i], j, distances[i]) {
+            if does_new_held_time_beat_record(u64::from(times[i]), j as u64, distances[i] as u64) {
                 times_beaten += 1;
             }
         }
         times_beaten *= 2;
 
-        if does_new_held_time_beat_record(times[i], different_tries, distances[i]) {
+        if does_new_held_time_beat_record(u64::from(times[i]), different_tries as u64, distances[i] as u64) {
             if times[i] % 2 == 0 {
                 times_beaten += 1;
             } else {
@@ -48,6 +48,6 @@ fn get_u16_numbers_from_string_with_colon(string: &str) -> Vec<u16> {
         .collect::<Vec<u16>>();
 }
 
-pub fn does_new_held_time_beat_record(total_time: u8, to_race_time: u8, goal_distance: u16) -> bool {
-    return (u16::from(total_time) - to_race_time as u16) * to_race_time as u16 > goal_distance;
+pub fn does_new_held_time_beat_record(total_time: u64, to_race_time: u64, goal_distance: u64) -> bool {
+    return (total_time - to_race_time) * to_race_time > goal_distance;
 }
